Support for while loops
- Broke out classes to files
- Introduced blocks to handle different endBlock instructions
